+++ Cargo.toml
@@ -112,6 +112,7 @@ members = [
     "sdk/gen-headers",
     "sdk/hash",
     "sdk/instruction",
+    "sdk/limited-deserialize",
     "sdk/macro",
     "sdk/msg",
     "sdk/package-metadata-macro",
@@ -409,6 +410,7 @@ solana-inline-spl = { path = "inline-spl", version = "=2.1.0" }
 solana-instruction = { path = "sdk/instruction", version = "=2.1.0", default-features = false }
 solana-lattice-hash = { path = "lattice-hash", version = "=2.1.0" }
 solana-ledger = { path = "ledger", version = "=2.1.0" }
+solana-limited-deserialize = { path = "sdk/limited-deserialize", version = "=2.1.0" }
 solana-loader-v4-program = { path = "programs/loader-v4", version = "=2.1.0" }
 solana-local-cluster = { path = "local-cluster", version = "=2.1.0" }
 solana-log-collector = { path = "log-collector", version = "=2.1.0" }
