+++ sdk/limited-deserialize/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "solana-limited-deserialize"
+description = "Solana function for deserializing with a limit"
+documentation = "https://docs.rs/solana-limited-deserialize"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bincode = { workspace = true }
+serde = { workspace = true }
+solana-instruction = { workspace = true, default-features = false, features = [
+    "std",
+] }
+
+[dev-dependencies]
+solana-program = { path = "../program" }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
